field change //Sachin

+++ src/main/java/com/ats/webapi/model/rawmaterial/ItemDetail.java
@@ -44,7 +44,7 @@ public class ItemDetail implements Serializable{
 	private int rmQty;
 	
 	@Column(name="rm_weight")
-	private int rmWeight;
+	private float rmWeight;
 	
 	@Column(name="del_status")
 	private int delStatus;
@@ -126,11 +126,12 @@ public class ItemDetail implements Serializable{
 		this.rmUomId = rmUomId;
 	}
 
-	public int getRmWeight() {
+
+	public float getRmWeight() {
 		return rmWeight;
 	}
 
-	public void setRmWeight(int rmWeight) {
+	public void setRmWeight(float rmWeight) {
 		this.rmWeight = rmWeight;
 	}
